Load and export (#156)
* Added some code for I/O

* Added better logic for FileUpload and added Blob configuration

* Got all tests to run correctly

* Removed previous  button

* Fixed test framework setup

* Fixed test framework setup

* Think its done boys

* Refactoring to reduce number of lines and files

* Fixed first level of testing

* Starting to pull down Develop testing

* Update class-area.component.css

Added css to anchor tag so it looks like the rest of his button friends

* Passing tests

Fixing some bugs that I found in testing. seems the model doesn't save the position or connections on a call to the classes by default. They now populate the array that is downloaded but are not currently applied to the imported yaml until returning from the CLI

* Added tests to analyze blob

* Added preliminary import testing